requirements.txt
# Core Data Science Libraries
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Machine Learning
scikit-learn==1.3.0

# NLP Libraries
nltk==3.8.1
spacy==3.6.0
textblob==0.17.1
vaderSentiment==3.3.2
nrclex==4.0
text2emotion==0.0.5

# Web Scraping & API
beautifulsoup4==4.12.2
newspaper3k==0.2.8
lxml==4.9.3
requests==2.31.0
newsapi-python==0.2.7

# Toxicity Detection
detoxify==0.5.1
torch==2.0.1
transformers==4.31.0

# Audio Transcription (optional)
assemblyai==0.17.0

# Web Framework
streamlit==1.25.0
flask==2.3.2

# Utilities
python-dotenv==1.0.0
wordcloud==1.9.2
Pillow==10.0.0

# PDF/Document Processing
PyPDF2==3.0.1
pdfplumber==0.9.0
python-docx==0.8.11

# Jupyter
jupyter==1.0.0
ipykernel==6.25.0
notebook==7.0.0
